+++ src/features/auth/index.ts
@@ -1,3 +1,4 @@
-export * from './types';
 export * from './hooks/useAuth';
+export * from './hooks/useAuthReturn';
 export * from './api/kakaoAuth';
+export * from './store/useAuthSessionStore';